desktop/requirements.txt
pywebview==5.4
requests==2.32.3

